+++ src/main.rs
@@ -1,6 +1,6 @@
 use prometheus::Registry;
 use psql_query_exporter::{
-    app_config::AppConfig, run_exporter, scrape_config::ScrapeConfig, utils::SignalHandler,
+    cli::CliParams, config::ScrapeConfig, run_exporter, utils::SignalHandler,
 };
 use std::{error::Error, net::SocketAddr};
 use tracing::instrument;
@@ -8,9 +8,9 @@ use tracing::instrument;
 #[tokio::main]
 #[instrument("Main")]
 async fn main() -> Result<(), Box<dyn Error>> {
-    let app_config = AppConfig::new();
-    let scrape_config = ScrapeConfig::from_file(&app_config.config)?;
-    let addr = SocketAddr::from((app_config.listen_on, app_config.port));
+    let cli = CliParams::new();
+    let scrape_config = ScrapeConfig::from_file(&cli.config)?;
+    let addr = SocketAddr::from((cli.listen_on, cli.port));
     let signal_handler = SignalHandler::new()?;
     let registry = Registry::new();
