+++ src/test_utils.rs
@@ -1,0 +1,19 @@
+use std::{
+    net::SocketAddr,
+    sync::atomic::{AtomicU16, Ordering},
+};
+use tokio::sync::OnceCell;
+
+pub fn next_addr() -> SocketAddr {
+    static PORT: AtomicU16 = AtomicU16::new(9000);
+
+    let next_port = PORT.fetch_add(1, Ordering::SeqCst);
+    format!("127.0.0.1:{next_port}").parse().unwrap()
+}
+
+pub async fn init_tracing() {
+    static INIT: OnceCell<()> = OnceCell::const_new();
+
+    INIT.get_or_init(async || tracing_subscriber::fmt::try_init().unwrap())
+        .await;
+}
